Move field commands into own namespace.
