+++ src/Drupal/Commands/core/drush.services.yml
@@ -21,50 +21,6 @@ services:
       arguments: ['@entity_type.manager']
       tags:
         -  { name: drush.command }
-  field.create.commands:
-    class: \Drush\Drupal\Commands\core\FieldCreateCommands
-    arguments:
-      - '@plugin.manager.field.field_type'
-      - '@plugin.manager.field.widget'
-      - '@plugin.manager.entity_reference_selection'
-      - '@entity_type.manager'
-      - '@entity_type.bundle.info'
-      - '@module_handler'
-      - '@entity_field.manager'
-    calls:
-      - [ setContentTranslationManager, [ '@?content_translation.manager' ] ]
-    tags:
-      -  { name: drush.command }
-  field.info.commands:
-    class: \Drush\Drupal\Commands\core\FieldInfoCommands
-    arguments:
-      - '@entity_type.manager'
-      - '@entity_type.bundle.info'
-    tags:
-      -  { name: drush.command }
-  field.delete.commands:
-    class: \Drush\Drupal\Commands\core\FieldDeleteCommands
-    arguments:
-      - '@entity_type.manager'
-      - '@entity_type.bundle.info'
-    tags:
-      -  { name: drush.command }
-  field.base-override-create.commands:
-    class: \Drush\Drupal\Commands\core\FieldBaseOverrideCreateCommands
-    arguments:
-      - '@entity_type.manager'
-      - '@entity_type.bundle.info'
-      - '@entity_field.manager'
-    tags:
-      - { name: drush.command }
-  field.base-info.commands:
-    class: \Drush\Drupal\Commands\core\FieldBaseInfoCommands
-    arguments:
-      - '@entity_type.manager'
-      - '@entity_type.bundle.info'
-      - '@entity_field.manager'
-    tags:
-      - { name: drush.command }
   link.hooks:
     class: \Drush\Drupal\Commands\core\LinkHooks
     arguments:
