+++ src/Drupal/Commands/field/EntityTypeBundleAskTrait.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace Drush\Drupal\Commands\core;
+namespace Drush\Drupal\Commands\field;
 
 use Drupal\Core\Entity\EntityTypeBundleInfoInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Symfony\Component\Console\Input\InputInterface;
+
+use function t;
 
 /**
  * @property InputInterface $input
