+++ src/Drupal/Commands/field/drush.services.yml
@@ -1,0 +1,45 @@
+services:
+  field.create.commands:
+    class: \Drush\Drupal\Commands\field\FieldCreateCommands
+    arguments:
+      - '@plugin.manager.field.field_type'
+      - '@plugin.manager.field.widget'
+      - '@plugin.manager.entity_reference_selection'
+      - '@entity_type.manager'
+      - '@entity_type.bundle.info'
+      - '@module_handler'
+      - '@entity_field.manager'
+    calls:
+      - [ setContentTranslationManager, [ '@?content_translation.manager' ] ]
+    tags:
+      -  { name: drush.command }
+    field.info.commands:
+      class: \Drush\Drupal\Commands\field\FieldInfoCommands
+      arguments:
+        - '@entity_type.manager'
+        - '@entity_type.bundle.info'
+      tags:
+        -  { name: drush.command }
+    field.delete.commands:
+      class: \Drush\Drupal\Commands\field\FieldDeleteCommands
+      arguments:
+        - '@entity_type.manager'
+        - '@entity_type.bundle.info'
+      tags:
+        -  { name: drush.command }
+    field.base-override-create.commands:
+      class: \Drush\Drupal\Commands\field\FieldBaseOverrideCreateCommands
+      arguments:
+        - '@entity_type.manager'
+        - '@entity_type.bundle.info'
+        - '@entity_field.manager'
+      tags:
+        - { name: drush.command }
+    field.base-info.commands:
+      class: \Drush\Drupal\Commands\field\FieldBaseInfoCommands
+      arguments:
+        - '@entity_type.manager'
+        - '@entity_type.bundle.info'
+        - '@entity_field.manager'
+      tags:
+        - { name: drush.command }
